+++ src/main/java/campustv/repository/DBNewRepository.java
@@ -6,30 +6,44 @@ import java.sql.Statement;
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Iterator;
+import java.util.LinkedList;
 import java.util.List;
 
 import campustv.dbutils.JDBCUtils;
 import campustv.dbutils.SQLUtils;
+import campustv.domain.Agenda;
+import campustv.domain.Content;
+import campustv.domain.Event;
 import campustv.domain.Image;
 import campustv.domain.New;
 import campustv.domain.Video;
+import campustv.repository.exceptions.PermissionException;
 import campustv.repository.exceptions.RepositoryException;
+import campustv.repository.exceptions.SessionException;
 
 public class DBNewRepository implements NewRepository {
 
-	public void addNew(New n) throws RepositoryException {	
+	public void addNew(New n, String accessToken) throws RepositoryException, SessionException, PermissionException {	
 		try{
-			ResultSet rs = DBRepositoryUtils.insertContent(n);
-			rs.next();
-			long id_content = rs.getLong(1);
-			n.setContentID(id_content);
-			DBRepositoryUtils.insertNew(n);
-			Iterator<Video> it =  n.getVideos();
-			if(it.hasNext())
-				DBRepositoryUtils.insertVideos(it, n);
-			Iterator<Image> it2 = n.getImages();
-			if(it2.hasNext())
-				DBRepositoryUtils.insertImages(it2, n);
+			if(DBRepositoryUtils.hasProducer(n.getProducerID())){
+				if(DBRepositoryUtils.hasLoggedIn(n.getProducerID(), accessToken)){
+					ResultSet rs = DBRepositoryUtils.insertContent(n);
+					rs.next();
+					long id_content = rs.getLong(1);
+					n.setContentID(id_content);
+					DBRepositoryUtils.insertNew(n);
+					Iterator<Video> it =  n.getVideos();
+					if(it.hasNext())
+						DBRepositoryUtils.insertVideos(it, n);
+					Iterator<Image> it2 = n.getImages();
+					if(it2.hasNext())
+						DBRepositoryUtils.insertImages(it2, n);
+				}
+				else
+					throw new SessionException("User not logged in");
+			}
+			else
+				throw new PermissionException("Don't have permissions");
 		} catch(SQLException e){
 			throw new RepositoryException(e.getMessage());
 			
@@ -46,7 +60,7 @@ public class DBNewRepository implements NewRepository {
 			Statement st = JDBCUtils.getStatement();
 			String query = "select * " +
 		             "from news, contents " +
-		             "where news.id_new = contents.id_content";
+		             "where news.id_new = contents.id_content order by contents.insertion_date ";
 			
 			ResultSet rs = JDBCUtils.executeQuery(st, query);
 			while(rs.next()){
@@ -110,11 +124,23 @@ public class DBNewRepository implements NewRepository {
 		}
 	}
 	
-	public void removeNew(long newID) throws RepositoryException {
+	public void removeNew(long newID, long producerID, String accessToken) throws RepositoryException, PermissionException, SessionException {
 		try{
-			Statement st = JDBCUtils.getStatement();
-			st.addBatch("delete from contents where id_content = " + newID);
-			st.executeBatch();
+			if(DBRepositoryUtils.hasProducer(producerID)){
+				if(DBRepositoryUtils.hasLoggedIn(producerID, accessToken)){
+					if(DBRepositoryUtils.hasOnwer(producerID, newID)){
+						Statement st = JDBCUtils.getStatement();
+						st.addBatch("delete from contents where id_content = " + newID);
+						st.executeBatch();
+					}
+					else
+						throw new PermissionException("User not owner of content");
+				}
+				else
+					throw new SessionException("User not logged in");
+			}
+			else
+				throw new PermissionException("Don't have permissions");
 		} catch(SQLException e){
 			throw new RepositoryException(e.getMessage());
 		}
@@ -122,7 +148,7 @@ public class DBNewRepository implements NewRepository {
 
 	@Override
 	public Iterator<Video> listVideosOfNews() throws RepositoryException {
-		List<Video> videos = new ArrayList<Video>();
+		List<Video> videos = new LinkedList<Video>();
 
 		try{
 			Statement st = JDBCUtils.getStatement();
@@ -161,7 +187,7 @@ public class DBNewRepository implements NewRepository {
 	@Override
 	public Iterator<Video> listVideosOfNew(long newID)
 			throws RepositoryException {
-		List<Video> videos = new ArrayList<Video>();
+		List<Video> videos = new LinkedList<Video>();
 
 		try{
 			Statement st = JDBCUtils.getStatement();
@@ -195,6 +221,89 @@ public class DBNewRepository implements NewRepository {
 			throw new RepositoryException(e.getMessage());
 		}
 		return videos.iterator();
+	}
+
+
+	public Iterator<New> getNewsByAuthor(String author)
+			throws RepositoryException {
+		List<New> news = new LinkedList<New>();
+
+		try{
+			Statement st = JDBCUtils.getStatement();
+			String query = "select * " +
+		             "from news, contents, producers, users " +
+		             "where news.id_new = contents.id_content and producers.id_producer = users.is_user and users.name = '" + author + "'";
+			
+			ResultSet rs = JDBCUtils.executeQuery(st, query);
+			while(rs.next()){
+				int newID = rs.getInt(1);
+				String title = rs.getString(2);;
+				@SuppressWarnings("unused")
+				int contentID = rs.getInt(3);
+				String insertion_date = SQLUtils.timestampToString(rs.getTimestamp(4));
+				String initial_issue = SQLUtils.timestampToString(rs.getTimestamp(5));
+				String final_issue = SQLUtils.timestampToString(rs.getTimestamp(6));
+				String description = rs.getString(7);
+				long producerID = rs.getLong(8);
+				String category = rs.getString(9);
+				String subcategory = rs.getString(10);
+				List<Image> images = DBRepositoryUtils.getImages(newID, new New());
+				List<Video> videos = DBRepositoryUtils.getVideos(newID, new New());
+				
+				New n = new New(newID, insertion_date, initial_issue, final_issue, 
+						description, category, subcategory, producerID, title, images, videos); 
+			    news.add(n);
+			}
+		}catch(SQLException e){
+			throw new RepositoryException(e.getMessage());
+		} catch (ParseException e) {
+			throw new RepositoryException(e.getMessage());
+		}
+		return news.iterator();
+	}
+
+	@Override
+	public Iterator<Agenda> getNewsDay(String scheduleDay)
+			throws RepositoryException {
+		List<Agenda> agendas = new LinkedList<Agenda>();
+
+		try{
+			Statement st = JDBCUtils.getStatement();
+			String query = "select * " +
+		             "from agenda, contents, news" +
+		             "where agenda.schedule_date = '" + scheduleDay + "' and agenda.id_content = contents.id_content and"
+		             		+ "agenda.id_content = news.id_news";
+			
+			ResultSet rs = JDBCUtils.executeQuery(st, query);
+			while(rs.next()){
+				long contentID = rs.getLong(1);
+				
+				String estimated_time = SQLUtils.timestampToString(rs.getTimestamp(3));
+				String time_transmission = SQLUtils.timestampToString(rs.getTimestamp(4));
+				
+				String insertion_date = SQLUtils.timestampToString(rs.getTimestamp(6));
+				String initial_issue = SQLUtils.timestampToString(rs.getTimestamp(7));
+				String final_issue = SQLUtils.timestampToString(rs.getTimestamp(8));
+				String description = rs.getString(9);
+				long producerID = rs.getLong(10);
+				String category = rs.getString(11);
+				String subcategory = rs.getString(12);
+				long newID = rs.getLong(13);
+				String title = rs.getString(14);;
+				List<Image> images = DBRepositoryUtils.getImages(newID, new New());
+				List<Video> videos = DBRepositoryUtils.getVideos(newID, new New());
+				Content c = new New(newID, insertion_date, initial_issue, final_issue, 
+						description, category, subcategory, producerID, title, images, videos); 
+				
+				Agenda a = new Agenda(c, scheduleDay, estimated_time, time_transmission);
+				agendas.add(a);
+			}
+		}catch(SQLException e){
+			throw new RepositoryException(e.getMessage());
+		} catch (ParseException e) {
+			throw new RepositoryException(e.getMessage());
+		}
+		return agendas.iterator();
 	}
